Updated simulation. Robot goes around twice and then exits
It goes around twice and then exits. It reads some parameters to a file,
and at the end outputs the total time it took to do 2 laps in
milliseconds to a file results.txt. It also has a penalty function added
to it if the game runts slow. Also if the robot runs off the track, then
it is detected and the output to results.txt is put very high as a
penelty function. We are going to optimize the turning  threshold,
ratio, speed, and updateSquareForAnalysis by linking the simulation to
VisualDoc and ModeFrontier.